requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask
opencv-python
image
Pillow
datetime
uuid
Flask-SQLAlchemy
PyJWT